requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.4
beautifultable==0.8.0
billiard==3.6.4.0
celery==4.4.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-aliases==1.0.1
click-default-group==1.2.2
click-didyoumean==0.0.3
click-help-colors==0.9
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
coverage==5.5
cryptography==3.4.8
Cython==0.29.23
da-vinci==0.3.0
Django==3.2
django-celery==3.3.1
django-crispy-forms==1.12.0
django-filter==2.4.0
django-thumbnails==0.6.0
djangorestframework==3.12.4
fandogh-cli==1.26.7
greenlet==1.1.0
idna==3.2
kombu==4.6.11
Markdown==3.3.4
Pillow==8.3.1
pkg_resources==0.0.0
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
pyc==0.1.0
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
redis-server==6.0.9
requests==2.26.0
requests-toolbelt==0.9.1
retrying==1.3.3
rich==9.13.0
shortuuid==1.0.1
six==1.16.0
sorl-thumbnail==12.7.0
SQLAlchemy==1.4.20
sqlparse==0.4.1
suds-py3==1.4.4.1
tabulate==0.8.9
task==0.2.5
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
vine==1.3.0
wcwidth==0.2.5
websocket-client==1.2.1
